Move command initialization into a file of its own.
